DAOS-12250 control: Log gRPC req before processing
The previous commit placed the request log behind a
conditional check in order to try to cut down on noisy
logging for requests that often result in a sentinel error
(e.g. ErrNotLeader). That approach does result in cleaner
logs, but has the downside of making requests more difficult
to trace, as the beginning of the request handling is not
logged first.

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>